+++ docs/docs/user-guide/changelog.md
@@ -13,12 +13,17 @@ sidebar_position: 5
 - Elevation controls
 - Add tooltips to variants panel
 - UI precision for metric slices returns to 0.1m. For more granular resolution of 0.05m, hold 'shift' when adjusting the width. This should be more obvious in the UI tooltip.
+- If street is not geolocated, map view will now center on user's current position if available
+
+### Improvements
+- Improve map view.
 
 ### 🐛 Bug fixes
 
 - Fix visual artifacts for the variants panel in right-to-left language modes.
 - Clearing a slice's custom label will now reset it to its original default label.
 - Fix a bug where after clearing a user-provided street name, changing its geo location will not automatically attach a new street name
+
 -->
 
 ## May 21, 2025